remove spotguides feature
