--- .gen/pipeline/pipeline/model_launch_spotguides_request.go
@@ -1,32 +1,0 @@
-/*
- * Pipeline API
- *
- * Pipeline is a feature rich application platform, built for containers on top of Kubernetes to automate the DevOps experience, continuous application development and the lifecycle of deployments. 
- *
- * API version: latest
- * Contact: info@banzaicloud.com
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package pipeline
-
-type LaunchSpotguidesRequest struct {
-
-	RepoOrganization string `json:"repoOrganization"`
-
-	RepoName string `json:"repoName"`
-
-	RepoPrivate bool `json:"repoPrivate,omitempty"`
-
-	RepoLatent bool `json:"repoLatent,omitempty"`
-
-	SpotguideName string `json:"spotguideName"`
-
-	SpotguideVersion string `json:"spotguideVersion,omitempty"`
-
-	Cluster CreateClusterRequest `json:"cluster,omitempty"`
-
-	Secrets []CreateSecretRequest `json:"secrets,omitempty"`
-
-	Pipeline map[string]interface{} `json:"pipeline,omitempty"`
-}
