--- .gen/pipeline/pipeline/model_spotguide_details_response.go
@@ -1,32 +1,0 @@
-/*
- * Pipeline API
- *
- * Pipeline is a feature rich application platform, built for containers on top of Kubernetes to automate the DevOps experience, continuous application development and the lifecycle of deployments. 
- *
- * API version: latest
- * Contact: info@banzaicloud.com
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package pipeline
-
-type SpotguideDetailsResponse struct {
-
-	Name string `json:"name,omitempty"`
-
-	DisplayName string `json:"displayName,omitempty"`
-
-	CreatedAt string `json:"createdAt,omitempty"`
-
-	UpdatedAt string `json:"updatedAt,omitempty"`
-
-	Readme string `json:"readme,omitempty"`
-
-	Version string `json:"version,omitempty"`
-
-	Tags []string `json:"tags,omitempty"`
-
-	Resources RequestedResources `json:"resources,omitempty"`
-
-	Questions []SpotguideOption `json:"questions,omitempty"`
-}
