--- .gen/pipeline/pipeline/model_spotguide_option.go
@@ -1,26 +1,0 @@
-/*
- * Pipeline API
- *
- * Pipeline is a feature rich application platform, built for containers on top of Kubernetes to automate the DevOps experience, continuous application development and the lifecycle of deployments. 
- *
- * API version: latest
- * Contact: info@banzaicloud.com
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package pipeline
-
-type SpotguideOption struct {
-
-	Name string `json:"name,omitempty"`
-
-	Type string `json:"type,omitempty"`
-
-	Info string `json:"info,omitempty"`
-
-	Default string `json:"default,omitempty"`
-
-	Readonly bool `json:"readonly,omitempty"`
-
-	Key string `json:"key,omitempty"`
-}
