+++ apis/pipeline/postman.json
@@ -4050,148 +4050,6 @@
 			"protocolProfileBehavior": {}
 		},
 		{
-			"name": "Spotguides",
-			"item": [
-				{
-					"name": "Spotguide List",
-					"request": {
-						"method": "GET",
-						"header": [],
-						"url": {
-							"raw": "{{url}}/api/v1/orgs/:orgId/spotguides",
-							"host": [
-								"{{url}}"
-							],
-							"path": [
-								"api",
-								"v1",
-								"orgs",
-								":orgId",
-								"spotguides"
-							],
-							"variable": [
-								{
-									"key": "orgId",
-									"value": "{{orgId}}",
-									"description": "Organization identifier"
-								}
-							]
-						}
-					},
-					"response": []
-				},
-				{
-					"name": "Spotguide Create",
-					"request": {
-						"method": "POST",
-						"header": [
-							{
-								"key": "Content-Type",
-								"value": "application/json"
-							}
-						],
-						"body": {
-							"mode": "raw",
-							"raw": "{\n\t\"spotguideName\": \"banzaicloud/spotguide-nodejs-mongodb\",\n\t\"repoOrganization\": \"banzaicloud\",\n\t\"repoName\": \"nodejs-spotguide-test\"\n}\n"
-						},
-						"url": {
-							"raw": "{{url}}/api/v1/orgs/:orgId/spotguides",
-							"host": [
-								"{{url}}"
-							],
-							"path": [
-								"api",
-								"v1",
-								"orgs",
-								":orgId",
-								"spotguides"
-							],
-							"variable": [
-								{
-									"key": "orgId",
-									"value": "{{orgId}}",
-									"description": "Organization identifier"
-								}
-							]
-						}
-					},
-					"response": []
-				},
-				{
-					"name": "Spotguide Sync Repos",
-					"request": {
-						"method": "PUT",
-						"header": [
-							{
-								"key": "Content-Type",
-								"value": "application/json"
-							}
-						],
-						"body": {
-							"mode": "raw",
-							"raw": ""
-						},
-						"url": {
-							"raw": "{{url}}/api/v1/orgs/:orgId/spotguides",
-							"host": [
-								"{{url}}"
-							],
-							"path": [
-								"api",
-								"v1",
-								"orgs",
-								":orgId",
-								"spotguides"
-							],
-							"variable": [
-								{
-									"key": "orgId",
-									"value": "{{orgId}}",
-									"description": "Organization identifier"
-								}
-							]
-						}
-					},
-					"response": []
-				},
-				{
-					"name": "Spotguide Details",
-					"request": {
-						"method": "GET",
-						"header": [],
-						"url": {
-							"raw": "{{url}}/api/v1/orgs/:orgId/spotguides/:spotguideName",
-							"host": [
-								"{{url}}"
-							],
-							"path": [
-								"api",
-								"v1",
-								"orgs",
-								":orgId",
-								"spotguides",
-								":spotguideName"
-							],
-							"variable": [
-								{
-									"key": "orgId",
-									"value": "{{orgId}}",
-									"description": "Organization identifier"
-								},
-								{
-									"key": "spotguideName",
-									"value": "banzaicloud/spotguide-nodejs-mongodb",
-									"description": "Spotguide name"
-								}
-							]
-						}
-					},
-					"response": []
-				}
-			],
-			"protocolProfileBehavior": {}
-		},
-		{
 			"name": "Deployments",
 			"item": [
 				{
