--- src/spotguide/log.go
@@ -1,32 +1,0 @@
-// Copyright © 2018 Banzai Cloud
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package spotguide
-
-import (
-	"github.com/sirupsen/logrus"
-
-	"github.com/banzaicloud/pipeline/internal/global"
-)
-
-// nolint: gochecknoglobals
-var log logrus.FieldLogger
-
-func init() {
-	log = global.LogrusLogger()
-
-	global.SubscribeLogrusLogger(func(l *logrus.Logger) {
-		log = l
-	})
-}
